requirements_datagen.txt
# FastRoute Data Generation Requirements
# ====================================

# Core data manipulation
pandas>=1.5.0
numpy>=1.21.0

# Database connectivity
mysql-connector-python>=8.0.33
SQLAlchemy>=1.4.0

# Data generation
Faker>=19.0.0
faker-vehicle>=0.2.0

# Geographic calculations
geopy>=2.3.0

# Progress tracking
tqdm>=4.64.0

# Concurrent processing (built-in in Python 3.2+)
# concurrent.futures

# Utilities
python-dateutil>=2.8.0
pytz>=2022.1

# Development
pytest>=7.0.0
pytest-mock>=3.8.0
